+++ src/main/kotlin/views/RoundView.kt
@@ -3,7 +3,7 @@ package com.github.juanalberticohf.views
 import com.github.juanalberticohf.models.Player
 import com.github.juanalberticohf.getPlayersTable
 import com.github.juanalberticohf.models.enums.PaymentMethod
-import com.github.juanalberticohf.models.input.InputRound
+import com.github.juanalberticohf.models.parsers.InputRound
 
 class RoundView {
     /**
